# Faster analysis with a big const array in a class