… tests/PHPStan/Analyser/data/bug-12159.php