closes https://github.com/phpstan/phpstan/issues/12159

before this PR
```
➜  phpstan-src git:(2.1.x) ✗ time php bin/phpstan analyze bug12159.php --debug 
....                                                       

php bin/phpstan analyze bug12159.php --debug  24.02s user 0.19s system 99% cpu 24.217 total
```

after this PR
```
➜  phpstan-src git:(perf) ✗ time php bin/phpstan analyze bug12159.php --debug 
...                                                                     
                                                                                                                        
php bin/phpstan analyze bug12159.php --debug  0.52s user 0.11s system 99% cpu 0.632 total
```